algo4: math: Stuffs from Leetcode

+++ algo4/math/count_trailing_zero_in_n_factorial.java
@@ -5,20 +5,24 @@ import java.io.PrintWriter;
 
 /**
  * Count the trailing zero in n!.
- * 
+ *
  * [Difficulty] - Medium
- * [Source]     - facebook interview, {@linkplain http://acm.zju.edu.cn/onlinejudge/showProblem.do?problemCode=2022}
+ * [Source]     - facebook interview,
+ *                {@linkplain http://acm.zju.edu.cn/onlinejudge/showProblem.do?problemCode=2022}
+ *                {@linkplain https://leetcode.com/problems/factorial-trailing-zeroes/}
  *
  */
-public class count_trailing_zero_in_n_factorial {
+public class count_trailing_zero_in_n_factorial
+{
     /**
-     * The key observation here is that how many 10 could be formed from the number
-     * 1 to n. 10 could be multiplied by 2 and 5, also because there will be an even
-     * number every 2 numbers, the task becomes counting how many number of 5 in all
-     * factors of n!.
-     * 
+     * The key observation here is that how many 10 could be formed from the
+     * number 1 to n. 10 could be multiplied by 2 and 5, also because there will
+     * be an even number every 2 numbers, the task becomes counting how many
+     * number of 5 in all factors of n!.
+     *
      */
-    public static int countTrailingZero(int n) {
+    public static int countTrailingZero(int n)
+    {
         int count = 0, divisor = 5;
 
         while (n / divisor > 0) {
@@ -34,17 +38,19 @@ public class count_trailing_zero_in_n_factorial {
         return count;
     }
 
-    public static int countTrailingZeroImproved(int n) {
+    public static int countTrailingZeroImproved(int n)
+    {
         int count = 0;
 
-        for (int divisor = 5; divisor <= n; divisor *= 5) {
+        for (long divisor = 5; divisor <= n; divisor *= 5) {
             count += n / divisor;
         }
 
         return count;
     }
 
-    public static void main(String[] args) throws NumberFormatException, IOException {
+    public static void main(String[] args) throws NumberFormatException, IOException
+    {
         BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
         PrintWriter out = new PrintWriter(System.out);
         int n = Integer.parseInt(in.readLine());
